% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy.profile.R
\name{canopy.profile}
\alias{canopy.profile}
\title{Mangrove Canopy Profile for Point-Centered Quarter Method}
\usage{
canopy.profile(x, samplingpoint = "samplingpoint", height = "height",
  interval10 = TRUE, interval = "interval", ymax = NULL, xmax = NULL)
}
\arguments{
\item{samplingpoint}{Column name for sampling points (numerical). Default name is "samplingpoint". First sampling point must be 1.}

\item{height}{Column name for height. Default name is "height". Values must be in meters.}

\item{interval10}{Logical argument for distance between PCQM sampling points. If interval10=TRUE (Default), the canopy plot assumes equidistant 10-meter spacing between all sampling points. If interval10=FALSE, a unique column must exist in the dataframe with distance from the previous sampling point.}

\item{interval}{Column name for the distances between sampling points (required if interval10=FALSE). Default name is "interval". The first row (samplingpoint 1) must be a value of 0. Each successive sampling point number represents the distance from the previous sampling point, NOT the additive distance from sampling point 1.}

\item{ymax}{Optional argument for specifying the maximum extent of the y axis. If not specified (Default), ymax is estimated based on dataframe values.}

\item{xmax}{Optional argument for specifying the maximum extent of the x axis. If not specified (Default), xmax is estimated based on dataframe values.}                                                                       
}
\description{
This function allows you to plot the canopy profile for mangrove forest transects from pcqm data.
}
\examples{
canopy.profile(mangrove_data)
canopy.profile(mangrove_data, ymax = 50, samplingpoint = "Sampling_Point", interval10=FALSE)
}
\keyword{canopy}
\keyword{mangrove}
\keyword{pcqm,}
\keyword{profile}
\keyword{structure,}
