% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.method.R
\name{circle.method}
\alias{circle.method}
\title{Mangrove Fixed-Radius Plot Method Analysis}
\usage{
plot.method(x, plotnumber = "plotnumber", dbh = "dbh",
  species = "species", height = "height", plot.radius = 5)
}
\arguments{
\item{plotnumber}{Column name in data frame for plot number (numerical). Default name is "plotnumber". First plot number must be 1.}

\item{dbh}{Column name in data frame for diameter at breast height. Default name is "dbh". Values must be in centimeters.}

\item{species}{Column name in data frame for species. Default name is "species".}

\item{height}{Column name in data frame for height (Optional). Default name is "height". Values must be in meters. When included, additional height-related outputs are displayed.}

\item{plot.radius}{Plot radius measured in meters. Default is 5 meters.}

}
\description{
This function allows you to estimate mangrove forest structure based on sampling using fixed-radius plot sampling.
}
\examples{
circle.method(mangrove_data)
circle.method(mangrove_data, species="Species", plot.radius=5, plotnumber = "Plot", dbh="Diameter", height="Tree.Height")
}
\keyword{mangrove}
\keyword{plot}
\keyword{structure,}
