% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.plot.R
\name{iv.plot}
\alias{iv.plot}
\title{Importance Value Radar Plot for Mangrove PCQM or Plot Data}
\usage{
iv.plot(x, colors = NULL)
}
\arguments{
\item{x}{Data frame object output from the pcqm.method() or plot.method() functions.}

\item{colors}{Optional list of colors to be used in the radar plot. Maximum of five species will be plotted. If listed, one color must be provided for each species. If not specified, default colors will be used (red, black, blue, orange, green).}

}
\description{
This function allows you to output a radar plot for objects created from pcqm.method() or plot.method().
}
\examples{
iv.plot(data_output)
iv.plot(data_output, colors = c("black" , "gray", "goldenrod1", "lightgreen", "red"))
}
\keyword{radarplot}
\keyword{mangrove}
\keyword{pcqm,}
\keyword{structure,}
Contact GitHub API Training Shop Blog About
