% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqm.indices.R
\name{pcqm.indices}
\alias{pcqm.indices}
\title{Mangrove Indices for Point-Centered Quarter Method}
\usage{
pcqm.indices(x, samplingpoint = "samplingpoint", dist = "dist",
  species = "species", dbh = "dbh", height = "height", sizebin = F)
}
\arguments{
\item{samplingpoint}{Column name in data frame for sampling points (numerical). Default name is "samplingpoint". First sampling point must be 1.}

\item{dist}{Column name in data frame for distance from tree to sampling point. Default name is "dist". Values must be in meters.}

\item{species}{Column name in data frame for species. Default name is "species".}

\item{dbh}{Column name in data frame for diameter at breast height. Default name is "dbh". Values must be in centimeters.}

\item{height}{Column name in data frame for height (Optional). Default name is "height". Values must be in meters. When included, additional height-related outputs are displayed.}

\item{sizebin}{Logical argument for displaying size class plot (Default is FALSE). If sizebin = TRUE, three graphs are generated: (1) an illustration of size class proportions: <5 cm, 5–10 cm, and >10 cm DBH; (2) an illustration of tree height size class proportions: <5, 5–10, and >10 m; and (3) a scatterplot of the relationship between DBH and height.}
}
\description{
This function allows you to calculate the Holdridge Complexity Index and Mean Stand Diameter for pcqm data.
}
\examples{
pcqm.indices(mangrove_data)
pcqm.indices(mangrove_data, dbh = "Diameter", height = "Tree_Height", samplingpoint = "Sampling_Point", sizebin = T)
}
\keyword{Complexity}
\keyword{Diameter}
\keyword{Holdridge}
\keyword{Index,}
\keyword{Mean}
\keyword{Stand}
\keyword{mangrove}
\keyword{pcqm,}
\keyword{structure,}
