% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqm.method.R
\name{pcqm.method}
\alias{pcqm.method}
\title{Mangrove Point-Centered Quarter Method Analysis}
\usage{
pcqm.method(x, samplingpoint = "samplingpoint", dist = "dist",
  species = "species", dbh = "dbh", height = "height")
}
\arguments{
\item{samplingpoint}{Column name in data frame for sampling points (numerical). Default name is "samplingpoint". First sampling point must be 1.}

\item{dist}{Column name in data frame for distance from tree to sampling point. Default name is "dist". Values must be in meters.}

\item{species}{Column name in data frame for species. Default name is "species".}

\item{dbh}{Column name in data frame for diameter at breast height. Default name is "dbh". Values must be in centimeters.}

\item{height}{Column name in data frame for height (Optional). Default name is "height". Values must be in meters. When included, additional height-related outputs are displayed.}

}
\description{
This function allows you to estimate mangrove forest structure based on sampling using the Point-Centered Quarter Method (PCQM). See Cottam and Curtis (1956) for informaiton on PCQM.
}
\examples{
pcqm.method(mangrove_data)
pcqm.method(mangrove_data, dbh = "Diameter", height = "Tree_Height", samplingpoint = "Sampling_Point")
}
\keyword{mangrove}
\keyword{pcqm}
\keyword{structure,}
