% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.indices.R
\name{plot.indices}
\alias{plot.indices}
\title{Mangrove Indices for Plot Method}
\usage{
plot.indices(x, plotnumber = "plotnumber", dbh = "dbh",
  species = "species", height = "height", plot.width = 10,
  plot.length = 10, ageplot = F)
}
\arguments{
\item{plotnumber}{Column name in dataframe for plot nymber (numerical). Default name is "plotnumber". First plot number must be 1.}

\item{dbh}{Column name in dataframe for diameter at breast height. Default name is "dbh". Values must be in centimeters.}

\item{species}{Column name in dataframe for species. Default name is "species".}

\item{height}{Column name in dataframe for height. Default name is "height". Values must be in meters.}

\item{plot.width}{Plot width measured in meters. Default is 10 meters.}

\item{plot.length}{Plot length measured in meters. Default is 10 meters.}

\item{ageplot}{Logical argument for displaying size class plot (Default is FALSE). If ageplot=TRUE, plot of size class proportions is displayed: seedling (<5 cm dbh), sapling (5-10 cm dbh), and pole (>10 cm dbh).}
}
\description{
This function allows you to calculate the Holdridge Complexity Index and Mean Stand Diameter for pcqm data.
}
\examples{
plot.indices(mangrove_data)
plot.indices(mangrove_data, plotnumber = "Plot", species = "Species", plot.width = 5, ageplot=T)
}
\keyword{Complexity}
\keyword{Diameter}
\keyword{Holdridge}
\keyword{Index,}
\keyword{Mean}
\keyword{Stand}
\keyword{mangrove}
\keyword{plot,}
\keyword{structure,}
