% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.method.R
\name{plot.method}
\alias{plot.method}
\title{Mangrove Plot Method Analysis}
\usage{
plot.method(x, plotnumber = "plotnumber", dbh = "dbh",
  species = "species", height = "height", plot.width = 10,
  plot.length = 10)
}
\arguments{
\item{plotnumber}{Column name in data frame for plot number (numerical). Default name is "plotnumber". First plot number must be 1.}

\item{dbh}{Column name in data frame for diameter at breast height. Default name is "dbh". Values must be in centimeters.}

\item{species}{Column name in data frame for species. Default name is "species".}

\item{height}{Column name in data frame for height (Optional). Default name is "height". Values must be in meters. When included, additional height-related outputs are displayed.}

\item{plot.width}{Plot width measured in meters. Default is 10 meters.}

\item{plot.length}{Plot length measured in meters. Default is 10 meters.}

}
\description{
This function allows you to estimate mangrove forest structure based on sampling using fixed-area sampling (plot).
}
\examples{
plot.method(mangrove_data)
plot.method(mangrove_data, species="Species", plot.width=5, plotnumber = "Plot", dbh="Diameter", height="Tree.Height")
}
\keyword{mangrove}
\keyword{plot}
\keyword{structure,}
