% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqm.method.R
\name{pcqm.method}
\alias{pcqm.method}
\title{Mangrove Point-Centered Quarter Method Analysis}
\usage{
pcqm.method(x, samplingpoint = "samplingpoint", dist = "dist",
  species = "species", dbh = "dbh", height = "height", ivplot = F)
}
\arguments{
\item{samplingpoint}{Column name in dataframe for sampling points (numerical). Default name is "samplingpoint". First sampling point must be 1.}

\item{dist}{Column name in dataframe for distance from tree to sampling point. Default name is "dist". Values must be in meters.}

\item{species}{Column name in dataframe for species. Default name is "species".}

\item{dbh}{Column name in dataframe for diameter at breast height. Default name is "dbh". Values must be in centimeters.}

\item{height}{Column name in dataframe for height. Default name is "height". Values must be in meters.}

\item{ivplot}{Logical argument for whether a radar plot of species importance values is displayed (Default is FALSE). If ivplot=TRUE, the top five species (ranked by importance values) will be plotted via radarchart() from package 'fmsb'.}
}
\description{
This function allows you to estimate mangrove forest structure based on sampling using the Point-Centered Quarter Method (PCQM). See Cottam and Curtis (1956) for informaiton on PCQM.
}
\examples{
pcqm.method(mangrove_data)
pcqm.method(mangrove_data, dbh = "Diameter", height = "Tree_Height", samplingpoint = "Sampling_Point", ivplot = T)
}
\keyword{mangrove}
\keyword{pcqm}
\keyword{structure,}
