% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.indices.R
\name{plot.indices}
\alias{plot.indices}
\title{Mangrove Indices for Plot Method}
\usage{
plot.indices(x, plotnumber = "plotnumber", dbh = "dbh",
  species = "species", height = "height", plot.width = 10,
  plot.length = 10, sizebin = F)
}
\arguments{
\item{plotnumber}{Column name in data frame for plot nymber (numerical). Default name is "plotnumber". First plot number must be 1.}

\item{dbh}{Column name in data frame for diameter at breast height. Default name is "dbh". Values must be in centimeters.}

\item{species}{Column name in data frame for species. Default name is "species".}

\item{height}{Column name in data frame for height (Optional). Default name is "height". Values must be in meters. When included, additional height-related outputs are displayed.}

\item{plot.width}{Plot width measured in meters. Default is 10 meters.}

\item{plot.length}{Plot length measured in meters. Default is 10 meters.}

\item{sizebin}{Logical argument for displaying size class plot (Default is FALSE). If sizebin = TRUE, three graphs are generated: (1) an illustration of diameter size class proportions: <5 cm, 5–10 cm, and >10 cm DBH; (2) an illustration of tree height size class proportions: <5, 5–10, and >10 m; and (3) a scatterplot of the relationship between DBH and height.}
}
\description{
This function allows you to calculate the Holdridge Complexity Index and Mean Stand Diameter for pcqm data.
}
\examples{
plot.indices(mangrove_data)
plot.indices(mangrove_data, plotnumber = "Plot", species = "Species", plot.width = 5, sizebin=T)
}
\keyword{Complexity}
\keyword{Diameter}
\keyword{Holdridge}
\keyword{Index,}
\keyword{Mean}
\keyword{Stand}
\keyword{mangrove}
\keyword{plot,}
\keyword{structure,}
