% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.method.R
\name{plot.method}
\alias{plot.method}
\title{Mangrove Plot Method Analysis}
\usage{
plot.method(x, plotnumber = "plotnumber", dbh = "dbh",
  species = "species", height = "height", plot.width = 10,
  plot.length = 10, ivplot = FALSE)
}
\arguments{
\item{plotnumber}{Column name in dataframe for plot nymber (numerical). Default name is "plotnumber". First plot number must be 1.}

\item{dbh}{Column name in dataframe for diameter at breast height. Default name is "dbh". Values must be in centimeters.}

\item{species}{Column name in dataframe for species. Default name is "species".}

\item{height}{Column name in dataframe for height. Default name is "height". Values must be in meters.}

\item{plot.width}{Plot width measured in meters. Default is 10 meters.}

\item{plot.length}{Plot length measured in meters. Default is 10 meters.}

\item{ivplot}{Logical argument for whether a radar plot of species importance values is displayed. If ivplot=TRUE, the top five species (ranked by importance values) will be plotted via radarchart() from package 'fmsb'.}
}
\description{
This function allows you to estimate mangrove forest structure based on sampling using the Point-Centered Quarter Method (PCQM). See Cottam and Curtis (1956) for informaiton on PCQM.
}
\examples{
plot.method(mangrove_data)
plot.method(mangrove_data, species="Species", plot.width=5, plotnumber = "Plot", dbh="Diameter", height="Tree.Height", ivplot=T)
}
\keyword{mangrove}
\keyword{plot}
\keyword{structure,}
